move ogParser code to utils module
