+++ src/browser_action/modules/utils.js
@@ -13,3 +13,18 @@ export function getImageWidth(imageUrl) {
     img.src = imageUrl;
   });
 }
+
+export function getOgParserCodeString() {
+  return `(function(){
+    const ogData = {};
+    const ogTags = document.querySelectorAll('meta[property^="og:"]');
+    ogTags.forEach(tag=>{
+      const key = tag.getAttribute("property").substr(3);
+      const value = tag.getAttribute("content");
+      if(key){
+        ogData[key] = value;
+      }
+    });
+    return ogData;
+  })()`;
+}
